Progress thus far with the checks if any items are present in the database (cannot call an iterative LINQ method if there are none present, apparently).

+++ StoreManager/StoreManager.Library/Database/DatabaseStorageRepository.cs
@@ -29,55 +29,79 @@ namespace StoreManager.Library.Database
         }
 
         public async Task<DataBundle> ReadAsync() {
-            return await Task.Run(() => new DataBundle());
+            DataBundle result = null;
+            return await Task.Run(() => result);
         }
 
         public async Task WriteAsync(DataBundle dataBundle) {
             await Task.Run(() => { });
         }
 
-        public async Task CreateSomeAsync<T>(List<SEntity> entities) where T : SEntity {
-            await _interfacerManager.CreateSomeAsync<T>(entities);
+        public async Task CreateManyAsync<T>(List<SEntity> entities) where T : SEntity {
+            await _interfacerManager.CreateManyAsync<T>(entities);
         }
 
         public async Task CreateOneAsync<T>(SEntity entity) where T : SEntity {
-            await _interfacerManager.CreateAsync<T>(entity);
+            await _interfacerManager.CreateOneAsync<T>(entity);
         }
 
         public async Task<List<T>> GetAllAsync<T>() where T : SEntity {
-            return await Task.Run(() => _interfacerManager.GetAllAsync<T>().Result.ConvertAll(t => t as T));
+            List<T> result = null;
+
+            if (_interfacerManager.AnyAsync<T>().Result) {
+                result = (await _interfacerManager.GetAllAsync<T>())
+                    .ConvertAll(t => t as T);
+            }
+
+            return await Task.Run(() => result ?? new List<T>());
         }
 
-        public async Task<List<T>> GetSomeAsync<T>(List<int> ids) where T : SEntity {
-            return await Task.Run(() => _interfacerManager.GetSomeAsync<T>(ids).Result.ConvertAll(t => t as T));
+        public async Task<List<T>> GetManyAsync<T>(List<int> ids) where T : SEntity {
+            List<T> result = null;
+
+            if (_interfacerManager.AnyAsync<T>().Result) {
+                result = (await _interfacerManager.GetSomeAsync<T>(ids))
+                    .ConvertAll(t => t as T);
+            }
+
+            return await Task.Run(() => result ?? new List<T>());
         }
 
         public async Task<T> GetOneAsync<T>(int id) where T : SEntity {
-            return await _interfacerManager.GetOneAsync<T>(id) as T;
+            T result = null;
+
+            if (_interfacerManager.AnyAsync<T>().Result)
+                result = await _interfacerManager.GetOneAsync<T>(id) as T;
+
+            return result;
         }
 
-        public async Task UpdateAllAsync<T>(List<SEntity> items) where T : SEntity {
-            await _interfacerManager.UpdateAllAsync<T>(items);
-        }
-
-        public async Task UpdateSomeAsync<T>(List<SEntity> items) where T : SEntity {
-            await _interfacerManager.UpdateSomeAsync<T>(items);
+        public async Task UpdateManyAsync<T>(List<SEntity> items) where T : SEntity {
+            if (_interfacerManager.AnyAsync<T>().Result) {
+                await _interfacerManager.UpdateManyAsync<T>(items);
+            } else {
+                await _interfacerManager.CreateManyAsync<T>(items);
+            }
         }
 
         public async Task UpdateOneAsync<T>(SEntity entity) where T : SEntity {
-            await _interfacerManager.UpdateOneAsync<T>(entity);
+            if (_interfacerManager.AnyAsync<T>().Result) {
+                await _interfacerManager.UpdateOneAsync<T>(entity);
+            } else {
+                await CreateOneAsync<T>(entity);
+            }
         }
 
-        public async Task DeleteAllAsync<T>() where T : SEntity {
-            await _interfacerManager.DeleteAllAsync<T>();
-        }
-
-        public async Task DeleteSomeAsync<T>(List<SEntity> entities) where T : SEntity {
-            await _interfacerManager.DeleteSomeAsync<T>(entities);
+        public async Task DeleteManyAsync<T>(List<SEntity> entities) where T : SEntity {
+            if (_interfacerManager.AnyAsync<T>().Result) {
+                await _interfacerManager.DeleteSomeAsync<T>(entities);
+            }
         }
 
         public async Task DeleteOneAsync<T>(SEntity entity) where T : SEntity {
-            await _interfacerManager.DeleteOneAsync<T>(entity);
+            if (_interfacerManager.AnyAsync<T>().Result) {
+                await _interfacerManager.DeleteOneAsync<T>(entity);
+            }
         }
     }
 }
